#450 move selections hook to map bplane
